requirements.txt
aiohttp>=3.9.0
websockets>=12.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.2.0
hmmlearn>=0.3.0
arch>=6.2.0
scipy>=1.12.0
python-telegram-bot>=21.0
pycryptodome>=3.20.0
